implement basic CLI for DIQS, some 64 bit fixes. To do: Get memory usage down when building DB from scratch. Uses ~10x what it should.

+++ src/persistance_layer/on_disk_persistance.d
@@ -103,7 +103,7 @@ class OnDiskPersistance : PersistanceLayer
 		m_add_jobs.insertBack(image);
 	}
 
-	size_t length() {
+	uint length() {
 		enforceEx!DatabaseDirtyException(!dirty);
 		return m_num_images;
 	}
